# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  handler.ContractAuditRequest:
    properties:
      auditComment:
        type: string
      auditStatus:
        type: string
      contractAmount:
        type: number
      contractImage:
        type: string
      contractName:
        type: string
      contractNumber:
        type: string
      contractNumberSuffix:
        type: string
      cooperationType:
        type: string
      cooperationYears:
        type: integer
      customerId:
        type: integer
      endDate:
        type: integer
      expiryHandlingStatus:
        type: string
      isOnline:
        type: boolean
      nodeCount:
        type: integer
      paymentAmount:
        type: number
      paymentImage:
        type: string
      paymentStatus:
        type: string
      remark:
        type: string
      serviceUserId:
        type: integer
      startDate:
        type: integer
      websiteName:
        type: string
      websiteUrl:
        type: string
      websiteUsername:
        type: string
    required:
    - auditStatus
    - contractName
    - contractNumber
    - customerId
    type: object
  handler.ContractCreateRequest:
    properties:
      auditStatus:
        type: string
      contractAmount:
        type: number
      contractImage:
        type: string
      contractName:
        type: string
      contractNumber:
        type: string
      contractNumberSuffix:
        type: string
      cooperationType:
        type: string
      cooperationYears:
        type: integer
      customerId:
        type: integer
      endDate:
        type: integer
      expiryHandlingStatus:
        type: string
      isOnline:
        type: boolean
      nodeCount:
        type: integer
      paymentAmount:
        type: number
      paymentImage:
        type: string
      paymentStatus:
        type: string
      remark:
        type: string
      serviceUserId:
        type: integer
      startDate:
        type: integer
      websiteName:
        type: string
      websiteUrl:
        type: string
      websiteUsername:
        type: string
    required:
    - contractName
    - contractNumber
    - customerId
    type: object
  handler.ContractUpdateRequest:
    properties:
      auditStatus:
        type: string
      contractAmount:
        type: number
      contractImage:
        type: string
      contractName:
        type: string
      contractNumber:
        type: string
      contractNumberSuffix:
        type: string
      cooperationType:
        type: string
      cooperationYears:
        type: integer
      customerId:
        type: integer
      endDate:
        type: integer
      expiryHandlingStatus:
        type: string
      isOnline:
        type: boolean
      nodeCount:
        type: integer
      paymentAmount:
        type: number
      paymentImage:
        type: string
      paymentStatus:
        type: string
      remark:
        type: string
      serviceUserId:
        type: integer
      startDate:
        type: integer
      websiteName:
        type: string
      websiteUrl:
        type: string
      websiteUsername:
        type: string
    required:
    - contractName
    - contractNumber
    - customerId
    type: object
...
  model.Contract:
    properties:
      auditComment:
        type: string
      auditStatus:
        type: string
      auditedAt:
        type: string
      auditedBy:
        type: integer
      auditedByName:
        type: string
      contractAmount:
        type: number
      contractImage:
        type: string
      contractName:
        type: string
      contractNumber:
        type: string
      cooperationType:
        type: string
      cooperationYears:
        type: integer
      createdAt:
        type: string
      customerId:
        type: integer
      customerName:
        type: string
      endDate:
        type: string
      expiryHandlingStatus:
        type: string
      id:
        type: integer
      isOnline:
        type: boolean
      nodeCount:
        type: integer
      paymentAmount:
        type: number
      paymentImage:
        type: string
      paymentStatus:
        type: string
      remark:
        type: string
      serviceUserId:
        type: integer
      serviceUserName:
        type: string
      startDate:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
      userName:
        type: string
      websiteName:
        type: string
      websiteUrl:
        type: string
      websiteUsername:
        type: string
    type: object
  model.ContractListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Contract'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 后端服务API文档
  title: Backend API
  version: "1.0"
paths:
...
  /api/v1/contracts:
    get:
      parameters:
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 付款状态
        in: query
        name: paymentStatus
        type: string
      - description: 合作类型
        in: query
        name: cooperationType
        type: string
      - description: 审核状态
        in: query
        name: auditStatus
        type: string
      - description: 到期处理状态
        in: query
        name: expiryHandlingStatus
        type: string
      - description: 签单人ID
        in: query
        name: userId
        type: integer
      - description: 客户ID
        in: query
        name: customerId
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ContractListResult'
              type: object
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取合同列表
      tags:
      - contracts
    post:
      consumes:
      - application/json
      parameters:
      - description: 合同信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ContractCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Contract'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "409":
          description: 请求冲突
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建合同
      tags:
      - contracts
  /api/v1/contracts/{id}:
    delete:
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除合同
      tags:
      - contracts
    get:
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Contract'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取合同详情
      tags:
      - contracts
    put:
      consumes:
      - application/json
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      - description: 合同信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ContractUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Contract'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新合同
      tags:
      - contracts
  /api/v1/contracts/{id}/audit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      - description: 审核信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ContractAuditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Contract'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 审核合同
      tags:
      - contracts
  /api/v1/contracts/check-number:
    get:
      parameters:
      - description: 合同编号
        in: query
        name: contractNumber
        required: true
        type: string
      - description: 排除的合同ID
        in: query
        name: excludeId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 校验合同编号是否可用
      tags:
      - contracts
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token, 格式: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"